% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nasa_ames_1001.R
\name{read_nasa_ames_1001}
\alias{read_nasa_ames_1001}
\title{Read NASA Ames 1001}
\usage{
read_nasa_ames_1001(file, NComm_header = FALSE, header_length = NULL, ...)
}
\arguments{
\item{file}{path to the file}

\item{NComm_header}{Logical default false. If the final line of normal comments contains the data header, set to true}

\item{header_length}{by default the header length is obtained from the first line of the file. Supply a value here to override this.}

\item{...}{additional arguments passed to read.table}
}
\description{
Reads a NASA Ames FFI 1001 file by removing the header and parsing the rest of the data using read.table()
}
\author{
W. S. Drysdale
}
